Restricted Xor / Not / Or to comparison or in expressions
To simplify the logic of building sparql queries out of filter
trees, the Xor / Not and Or logical operators allow only
comparison or in expressions.  This behaviour may be revised
later.

The change also discriminates into separate operation types
comparison expressions and 'in' expressions.  It allows expressing
the logic on the filter trees easily.

TermCollection is no longer a Term type.