Implemented a filtering AST with a json decoder definition (#3)
* Implemented a filtering AST with a json decoder definition

* Fixed linting and jena iniatialization failures

* Restricted Xor / Not / Or to comparison or in expressions

To simplify the logic of building sparql queries out of filter
trees, the Xor / Not and Or logical operators allow only
comparison or in expressions.  This behaviour may be revised
later.

The change also discriminates into separate operation types
comparison expressions and 'in' expressions.  It allows expressing
the logic on the filter trees easily.

TermCollection is no longer a Term type.

* Added FilteredQuery to serialize filters into SPARQL queries

The functions in FilteredQuery unwrap the Filter AST into a
SPARQL query that also computes the total result count and adds
the pagination offset and limit.

The change also corrects the lexical form representation for
literal types when building the filter ast.

* Updated instance listing to make use of the generic filters

Introduced an new filter expression NoopExpr that allows selecting
all triples in the index.

* Added entity navigation route with target filtering

Added '/outgoing' as an instance sub-resource to list instances
that are linked to this instance through an outgoing link from
this instance.

The route accepts a filter query parameter that is applied on the
collection of outgoing instances.

* Added entity navigation route with source filtering

Added '/incoming' as an instance sub-resource to list instances
that are linked to this instance through an incoming link towards
this instance.

The route accepts a filter query parameter that is applied on the
collection of source instances.

* Fixed scapegoat error, added IncorrectlyNamedExceptions exception

* Fixed FilteredQuerySpec test failure